requirements-canva.txt
# 10x Canva Skills - Python Dependencies
# Install with: pip install -r requirements.txt

# ==========================================
# CANVA API DEPENDENCIES
# ==========================================

# HTTP requests for Canva API
requests>=2.28.0

# Environment variable management
python-dotenv>=1.0.0

# CLI formatting
rich>=13.0.0

# ==========================================
# LOCAL FILE EDITING DEPENDENCIES
# ==========================================

# PDF manipulation
PyPDF2>=3.0.0
pdfplumber>=0.9.0
reportlab>=4.0.0
pdf2docx>=0.5.6

# PowerPoint manipulation
python-pptx>=0.6.21
lxml>=4.9.0  # Required for slide duplication with theme preservation

# Word document manipulation
python-docx>=0.8.11

# Excel manipulation
openpyxl>=3.1.0
pandas>=2.0.0

# Image processing
Pillow>=9.0.0

# ==========================================
# OPTIONAL DEPENDENCIES
# ==========================================

# For local OAuth server
flask>=2.0.0

# For async operations
aiohttp>=3.8.0

# For better JSON handling
orjson>=3.8.0

# For HTML conversion
mammoth>=1.6.0

# For Markdown processing
markdown>=3.4.0

# ==========================================
# WINDOWS-SPECIFIC (for PDF conversion)
# ==========================================
# Uncomment if on Windows:
# comtypes>=1.2.0

# ==========================================
# DEVELOPMENT DEPENDENCIES
# ==========================================
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0
